Fix timer lifecycle: stop auto-lock and auto-backup timers on lock/quit
Extracts the auto-backup timer into src/main/auto-backup-timer.ts to allow
auto-lock.ts to import it without a circular dependency, then wires up the
missing stop/start calls:

- #222 / #179 / #157: AutoLockManager.lock() now calls stopAutoBackupTimer()
  before closing the DB, preventing the hourly backup tick from firing against
  a closed DB handle after a lock. startAutoBackupTimer() now stops any
  existing interval before creating a new one (previously used an early-return
  guard), so each unlock always gets a fresh, correctly-timed interval instead
  of inheriting a stale one from before the last lock.
- #272: before-quit handler now calls autoLock.stop() and stopAutoBackupTimer()
  before starting the async runAutoBackup(), so the idle-check and backup-tick
  intervals cannot fire concurrently with the shutdown backup.
- window-all-closed also calls stopAutoBackupTimer() for completeness.
